requirements.txt
# Core
streamlit==1.42.0
plotly==5.24.1
pandas==2.2.3
numpy==1.26.4
akshare>=1.16.72
openai==1.70.0
python-dotenv==1.0.1
requests>=2.31.0

# v5.0 Quantum Pro
redis==5.2.1
celery==5.4.0
bcrypt==4.2.1
pyjwt==2.10.1
sqlalchemy==2.0.37
google-generativeai==0.8.4
anthropic==0.44.0
scikit-learn==1.6.1
yfinance==0.2.55

# Testing (dev only)
pytest==8.3.4
